nsyllable methods for tokens

Extends nsyllable() methods for tokens objects.

# S3 method for tokens
nsyllable(
  x,
  language = "en",
  syllable_dictionary = nsyllable::data_syllables_en,
  use.names = FALSE
)

Arguments

x

character vector whose syllables will be counted. This will count all syllables in a character vector without regard to separating tokens, so it is recommended that x be individual terms.

language

specify the language for syllable counts by ISO 639-1 code. The default is English, using the data object data_syllables_en, an English pronunciation dictionary from CMU.

syllable_dictionary

optional named integer vector of syllable counts where the names are lower case tokens. This can be used to override the language setting, when set to NULL (the default). If a syllable dictionary is supplied, this will override the language argument.

use.names

logical; if TRUE, assign the tokens as the names of the syllable count vector

Examples

# \dontshow{
library("nsyllable")
txt <- c(one = "super freakily yes",
         two = "merrily all go aerodynamic")
toks <- quanteda::tokens(txt)
nsyllable(toks)
#> $one
#> [1] 2 3 1
#> 
#> $two
#> [1] 3 1 1 5
#> 
# }