Calculate keyness statistics

Calculate "keyness", a score for features that occur differentially across different categories. Here, the categories are defined by reference to a "target" document index in the dfm, with the reference group consisting of all other documents.

textstat_keyness(
  x,
  target = 1L,
  measure = c("chi2", "exact", "lr", "pmi"),
  sort = TRUE,
  correction = c("default", "yates", "williams", "none"),
  ...
)

Arguments

x

a dfm containing the features to be examined for keyness

target

the document index (numeric, character or logical) identifying the document forming the "target" for computing keyness; all other documents' feature frequencies will be combined for use as a reference

measure

(signed) association measure to be used for computing keyness. Currently available: "chi2"; "exact" (Fisher's exact test); "lr" for the likelihood ratio; "pmi" for pointwise mutual information. Note that the "exact" test is very computationally intensive and therefore much slower than the other methods.

sort

logical; if TRUE sort features scored in descending order of the measure, otherwise leave in original feature order

correction

if "default", Yates correction is applied to "chi2"; William's correction is applied to "lr"; and no correction is applied for the "exact" and "pmi" measures. Specifying a value other than the default can be used to override the defaults, for instance to apply the Williams correction to the chi2 measure. Specifying a correction for the "exact" and "pmi" measures has no effect and produces a warning.

...

not used

Value

a data.frame of computed statistics and associated p-values, where the features scored name each row, and the number of occurrences for both the target and reference groups. For measure = "chi2" this is the chi-squared value, signed positively if the observed value in the target exceeds its expected value; for measure = "exact" this is the estimate of the odds ratio; for measure = "lr" this is the likelihood ratio \(G2\) statistic; for "pmi" this is the pointwise mutual information statistics.

textstat_keyness returns a data.frame of features and their keyness scores and frequency counts.

References

Bondi, M. & Scott, M. (eds) (2010). Keyness in Texts. Amsterdam, Philadelphia: John Benjamins.

Stubbs, M. (2010). Three Concepts of Keywords. In Keyness in Texts, Bondi, M. & Scott, M. (eds): 1--42. Amsterdam, Philadelphia: John Benjamins.

Scott, M. & Tribble, C. (2006). Textual Patterns: Keyword and Corpus Analysis in Language Education. Amsterdam: Benjamins: 55.

Dunning, T. (1993). Accurate Methods for the Statistics of Surprise and Coincidence. Computational Linguistics, 19(1): 61--74.

Examples

library("quanteda")

# compare pre- v. post-war terms using grouping
period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
dfmat1 <- tokens(data_corpus_inaugural) %>%
    dfm() %>%
    dfm_group(groups = period)
head(dfmat1) # make sure 'post-war' is in the first row
#> Document-feature matrix of: 2 documents, 9,437 features (34.79% sparse) and 0 docvars.
#>           features
#> docs       fellow-citizens   of  the senate  and house representatives   :
#>   post-war               0 1514 2089      2 1552     3               0 115
#>   pre-war               39 5666 8094     13 3854     8              19  29
#>           features
#> docs       among vicissitudes
#>   post-war    22            0
#>   pre-war     86            5
#> [ reached max_nfeat ... 9,427 more features ]
head(tstat1 <- textstat_keyness(dfmat1), 10)
#>      feature     chi2 p n_target n_reference
#> 1         we 764.4484 0     1048         779
#> 2          . 300.1887 0     2014        3141
#> 3         us 207.5181 0      289         216
#> 4          - 205.0408 0      242         157
#> 5    america 200.5544 0      148          54
#> 6          : 187.9821 0      115          29
#> 7        our 183.5362 0      917        1307
#> 8      world 171.9660 0      196         123
#> 9  americans 163.1532 0       76           7
#> 10     today 137.7753 0       84          21
tail(tstat1, 10)
#>           feature       chi2            p n_target n_reference
#> 9428         upon  -58.39810 2.142730e-14       39         332
#> 9429       public  -58.87042 1.687539e-14       12         213
#> 9430 constitution  -59.66134 1.121325e-14        9         200
#> 9431           it  -60.68425 6.661338e-15      266        1132
#> 9432       states  -63.87803 1.332268e-15       29         305
#> 9433           be  -72.68161 0.000000e+00      278        1224
#> 9434       should  -88.14817 0.000000e+00       16         309
#> 9435        which -177.10087 0.000000e+00       96         911
#> 9436           of -197.08830 0.000000e+00     1514        5666
#> 9437          the -331.99069 0.000000e+00     2089        8094

# compare pre- v. post-war terms using logical vector
dfmat2 <- dfm(tokens(data_corpus_inaugural))
head(textstat_keyness(dfmat2, docvars(data_corpus_inaugural, "Year") >= 1945), 10)
#>      feature     chi2 p n_target n_reference
#> 1         we 764.4484 0     1048         779
#> 2          . 300.1887 0     2014        3141
#> 3         us 207.5181 0      289         216
#> 4          - 205.0408 0      242         157
#> 5    america 200.5544 0      148          54
#> 6          : 187.9821 0      115          29
#> 7        our 183.5362 0      917        1307
#> 8      world 171.9660 0      196         123
#> 9  americans 163.1532 0       76           7
#> 10     today 137.7753 0       84          21

# compare Trump 2017 to other post-war preseidents
dfmat3 <- dfm(tokens(corpus_subset(data_corpus_inaugural, period == "post-war")))
head(textstat_keyness(dfmat3, target = "2017-Trump"), 10)
#>      feature     chi2            p n_target n_reference
#> 1  protected 81.83024 0.000000e+00        5           1
#> 2      while 51.79484 6.161738e-13        6           7
#> 3      obama 51.05861 8.965051e-13        3           0
#> 4      we've 51.05861 8.965051e-13        3           0
#> 5       will 48.03251 4.192091e-12       40         332
#> 6   everyone 29.76164 4.885651e-08        4           5
#> 7       your 28.60175 8.890179e-08       11          51
#> 8    america 27.57968 1.507539e-07       18         130
#> 9     breath 27.27421 1.765507e-07        2           0
#> 10    exists 27.27421 1.765507e-07        2           0

# using the likelihood ratio method
head(textstat_keyness(dfm_smooth(dfmat3), measure = "lr", target = "2017-Trump"), 10)
#>      feature        G2            p n_target n_reference
#> 1       will 22.609878 1.984616e-06       41         351
#> 2    america 12.306921 4.512817e-04       19         149
#> 3       your 10.868622 9.780727e-04       12          70
#> 4      while  9.707425 1.835249e-03        7          26
#> 5      again  9.345219 2.235679e-03       10          56
#> 6  protected  8.909125 2.837491e-03        6          20
#> 7   american  7.996610 4.686501e-03       12          86
#> 8       back  7.113978 7.648521e-03        7          35
#> 9     dreams  5.908744 1.506591e-02        6          30
#> 10   country  5.725811 1.671732e-02       10          77