Select rows of textstat objects by glob, regex or fixed patterns

Users can subset output object of textstat_collocations, textstat_keyness or textstat_frequency based on "glob", "regex" or "fixed" patterns using this method.

textstat_select(
  x,
  pattern = NULL,
  selection = c("keep", "remove"),
  valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE
)

Arguments

x

a textstat object

pattern

see quanteda::pattern

selection

whether to "keep" or "remove" the rows that match the pattern

valuetype

the type of pattern matching: "glob" for "glob"-style wildcard expressions; "regex" for regular expressions; or "fixed" for exact matching. See valuetype for details.

case_insensitive

logical; if TRUE, ignore case when matching a pattern or dictionary values

Examples

library("quanteda")

period <- ifelse(docvars(data_corpus_inaugural, "Year") < 1945, "pre-war", "post-war")
dfmat <- tokens(data_corpus_inaugural) %>%
    dfm() %>%
    dfm_group(groups = period)
tstat <- textstat_keyness(dfmat)
textstat_select(tstat, 'america*')
#>          feature        chi2            p n_target n_reference
#> 5        america 200.5543560 0.000000e+00      148          54
#> 9      americans 163.1532091 0.000000e+00       76           7
#> 17     america's  93.4124870 0.000000e+00       37           0
#> 86      american  24.4057333 7.803611e-07       78          94
#> 1127    americas   0.6901897 4.060998e-01        2           1
#> 1698  american's   0.2300602 6.314792e-01        1           0
#> 5393 americanism  -0.3961920 5.290624e-01        0           1