Similarity and distance computation between documents or features

These functions compute matrixes of distances and similarities between documents or features from a dfm() and return a matrix of similarities or distances in a sparse format. These methods are fast and robust because they operate directly on the sparse dfm objects. The output can easily be coerced to an ordinary matrix, a data.frame of pairwise comparisons, or a dist format.

textstat_simil(
  x,
  y = NULL,
  selection = NULL,
  margin = c("documents", "features"),
  method = c("correlation", "cosine", "jaccard", "ejaccard", "dice", "edice", "hamann",
    "simple matching"),
  min_simil = NULL,
  ...
)

textstat_dist(
  x,
  y = NULL,
  selection = NULL,
  margin = c("documents", "features"),
  method = c("euclidean", "manhattan", "maximum", "canberra", "minkowski"),
  p = 2,
  ...
)

Arguments

x, y

a dfm objects; y is an optional target matrix matching x in the margin on which the similarity or distance will be computed.

selection

(deprecated - use y instead).

margin

identifies the margin of the dfm on which similarity or difference will be computed: "documents" for documents or "features" for word/term features.

method

character; the method identifying the similarity or distance measure to be used; see Details.

min_simil

numeric; a threshold for the similarity values below which similarity values will not be returned

...

unused

p

The power of the Minkowski distance.

Value

A sparse matrix from the Matrix package that will be symmetric unless y is specified.

Details

textstat_simil options are: "correlation" (default), "cosine", "jaccard", "ejaccard", "dice", "edice", "simple matching", and "hamann".

textstat_dist options are: "euclidean" (default), "manhattan", "maximum", "canberra", and "minkowski".

Note

If you want to compute similarity on a "normalized" dfm object (controlling for variable document lengths, for methods such as correlation for which different document lengths matter), then wrap the input dfm in [dfm_weight](x, "prop").

Conversion to other data types

The output objects from textstat_simil() and textstat_dist() can be transformed easily into a list format using as.list(), which returns a list for each unique element of the second of the pairs, a data.frame using as.data.frame(), which returns pairwise scores, as.dist()for a dist object, or as.matrix() to convert it into an ordinary matrix.

See also

Examples

# similarities for documents
library("quanteda")
dfmat <- corpus_subset(data_corpus_inaugural, Year > 2000) %>%
    tokens(remove_punct = TRUE) %>%
    tokens_remove(stopwords("english")) %>%
    dfm()
(tstat1 <- textstat_simil(dfmat, method = "cosine", margin = "documents"))
#> textstat_simil object; method = "cosine"
#>            2001-Bush 2005-Bush 2009-Obama 2013-Obama 2017-Trump 2021-Biden
#> 2001-Bush      1.000     0.520      0.541      0.556      0.452      0.562
#> 2005-Bush      0.520     1.000      0.458      0.516      0.435      0.480
#> 2009-Obama     0.541     0.458      1.000      0.637      0.448      0.616
#> 2013-Obama     0.556     0.516      0.637      1.000      0.455      0.606
#> 2017-Trump     0.452     0.435      0.448      0.455      1.000      0.513
#> 2021-Biden     0.562     0.480      0.616      0.606      0.513      1.000
as.matrix(tstat1)
#>            2001-Bush 2005-Bush 2009-Obama 2013-Obama 2017-Trump 2021-Biden
#> 2001-Bush  1.0000000 0.5204355  0.5411649  0.5561972  0.4518935  0.5619136
#> 2005-Bush  0.5204355 1.0000000  0.4575297  0.5163644  0.4349030  0.4797651
#> 2009-Obama 0.5411649 0.4575297  1.0000000  0.6373318  0.4481950  0.6158540
#> 2013-Obama 0.5561972 0.5163644  0.6373318  1.0000000  0.4546945  0.6061256
#> 2017-Trump 0.4518935 0.4349030  0.4481950  0.4546945  1.0000000  0.5133378
#> 2021-Biden 0.5619136 0.4797651  0.6158540  0.6061256  0.5133378  1.0000000
as.list(tstat1)
#> $`2001-Bush`
#> 2021-Biden 2013-Obama 2009-Obama  2005-Bush 2017-Trump 
#>  0.5619136  0.5561972  0.5411649  0.5204355  0.4518935 
#> 
#> $`2005-Bush`
#>  2001-Bush 2013-Obama 2021-Biden 2009-Obama 2017-Trump 
#>  0.5204355  0.5163644  0.4797651  0.4575297  0.4349030 
#> 
#> $`2009-Obama`
#> 2013-Obama 2021-Biden  2001-Bush  2005-Bush 2017-Trump 
#>  0.6373318  0.6158540  0.5411649  0.4575297  0.4481950 
#> 
#> $`2013-Obama`
#> 2009-Obama 2021-Biden  2001-Bush  2005-Bush 2017-Trump 
#>  0.6373318  0.6061256  0.5561972  0.5163644  0.4546945 
#> 
#> $`2017-Trump`
#> 2021-Biden 2013-Obama  2001-Bush 2009-Obama  2005-Bush 
#>  0.5133378  0.4546945  0.4518935  0.4481950  0.4349030 
#> 
#> $`2021-Biden`
#> 2009-Obama 2013-Obama  2001-Bush 2017-Trump  2005-Bush 
#>  0.6158540  0.6061256  0.5619136  0.5133378  0.4797651 
#> 
as.list(tstat1, diag = TRUE)
#> $`2001-Bush`
#>  2001-Bush 2021-Biden 2013-Obama 2009-Obama  2005-Bush 2017-Trump 
#>  1.0000000  0.5619136  0.5561972  0.5411649  0.5204355  0.4518935 
#> 
#> $`2005-Bush`
#>  2005-Bush  2001-Bush 2013-Obama 2021-Biden 2009-Obama 2017-Trump 
#>  1.0000000  0.5204355  0.5163644  0.4797651  0.4575297  0.4349030 
#> 
#> $`2009-Obama`
#> 2009-Obama 2013-Obama 2021-Biden  2001-Bush  2005-Bush 2017-Trump 
#>  1.0000000  0.6373318  0.6158540  0.5411649  0.4575297  0.4481950 
#> 
#> $`2013-Obama`
#> 2013-Obama 2009-Obama 2021-Biden  2001-Bush  2005-Bush 2017-Trump 
#>  1.0000000  0.6373318  0.6061256  0.5561972  0.5163644  0.4546945 
#> 
#> $`2017-Trump`
#> 2017-Trump 2021-Biden 2013-Obama  2001-Bush 2009-Obama  2005-Bush 
#>  1.0000000  0.5133378  0.4546945  0.4518935  0.4481950  0.4349030 
#> 
#> $`2021-Biden`
#> 2021-Biden 2009-Obama 2013-Obama  2001-Bush 2017-Trump  2005-Bush 
#>  1.0000000  0.6158540  0.6061256  0.5619136  0.5133378  0.4797651 
#> 

# min_simil
(tstat2 <- textstat_simil(dfmat, method = "cosine", margin = "documents", min_simil = 0.6))
#> textstat_simil object; method = "cosine"
#>            2001-Bush 2005-Bush 2009-Obama 2013-Obama 2017-Trump 2021-Biden
#> 2001-Bush          1         .          .          .          .          .
#> 2005-Bush          .         1          .          .          .          .
#> 2009-Obama         .         .      1.000      0.637          .      0.616
#> 2013-Obama         .         .      0.637      1.000          .      0.606
#> 2017-Trump         .         .          .          .          1          .
#> 2021-Biden         .         .      0.616      0.606          .      1.000
as.matrix(tstat2)
#>            2001-Bush 2005-Bush 2009-Obama 2013-Obama 2017-Trump 2021-Biden
#> 2001-Bush          1        NA         NA         NA         NA         NA
#> 2005-Bush         NA         1         NA         NA         NA         NA
#> 2009-Obama        NA        NA  1.0000000  0.6373318         NA  0.6158540
#> 2013-Obama        NA        NA  0.6373318  1.0000000         NA  0.6061256
#> 2017-Trump        NA        NA         NA         NA          1         NA
#> 2021-Biden        NA        NA  0.6158540  0.6061256         NA  1.0000000

# similarities for for specific documents
textstat_simil(dfmat, dfmat["2017-Trump", ], margin = "documents")
#> textstat_simil object; method = "correlation"
#>            2017-Trump
#> 2001-Bush       0.375
#> 2005-Bush       0.355
#> 2009-Obama      0.356
#> 2013-Obama      0.373
#> 2017-Trump      1.000
#> 2021-Biden      0.449
textstat_simil(dfmat, dfmat["2017-Trump", ], method = "cosine", margin = "documents")
#> textstat_simil object; method = "cosine"
#>            2017-Trump
#> 2001-Bush       0.452
#> 2005-Bush       0.435
#> 2009-Obama      0.448
#> 2013-Obama      0.455
#> 2017-Trump      1.000
#> 2021-Biden      0.513
textstat_simil(dfmat, dfmat[c("2009-Obama", "2013-Obama"), ], margin = "documents")
#> textstat_simil object; method = "correlation"
#>            2009-Obama 2013-Obama
#> 2001-Bush       0.452      0.479
#> 2005-Bush       0.352      0.432
#> 2009-Obama      1.000      0.561
#> 2013-Obama      0.561      1.000
#> 2017-Trump      0.356      0.373
#> 2021-Biden      0.548      0.543

# compute some term similarities
tstat3 <- textstat_simil(dfmat, dfmat[, c("fair", "health", "terror")], method = "cosine",
                         margin = "features")
head(as.matrix(tstat3), 10)
#>                    fair    health     terror
#> president     0.3396831 0.6240377 0.09805807
#> clinton       0.4714045 0.4330127 0.00000000
#> distinguished 0.5773503 0.7071068 0.00000000
#> guests        0.5773503 0.7071068 0.00000000
#> fellow        0.4256283 0.7298004 0.14744196
#> citizens      0.7064173 0.6488857 0.07647191
#> peaceful      0.5163978 0.6324555 0.00000000
#> transfer      0.3333333 0.4082483 0.00000000
#> authority     0.8164966 0.5000000 0.00000000
#> rare          0.4082483 0.5000000 0.00000000
as.list(tstat3, n = 6)
#> $fair
#>    continue      chance differences     dangers      choose     charity 
#>           1           1           1           1           1           1 
#> 
#> $health
#> generations     without        work      common     fathers      nation 
#>   0.9733285   0.9594032   0.9527861   0.9486833   0.9486833   0.9282422 
#> 
#> $terror
#>    bestowed  sacrifices   ancestors  generosity cooperation  forty-four 
#>           1           1           1           1           1           1 
#> 

# distances for documents
(tstat4 <- textstat_dist(dfmat, margin = "documents"))
#> textstat_dist object; method = "euclidean"
#>            2001-Bush 2005-Bush 2009-Obama 2013-Obama 2017-Trump 2021-Biden
#> 2001-Bush          0      52.8       49.9       48.3       47.6       57.2
#> 2005-Bush       52.8         0       60.8       56.9       57.4       66.0
#> 2009-Obama      49.9      60.8          0       48.0       54.9       56.1
#> 2013-Obama      48.3      56.9       48.0          0       53.7       56.5
#> 2017-Trump      47.6      57.4       54.9       53.7          0       60.0
#> 2021-Biden      57.2      66.0       56.1       56.5       60.0          0
as.matrix(tstat4)
#>            2001-Bush 2005-Bush 2009-Obama 2013-Obama 2017-Trump 2021-Biden
#> 2001-Bush    0.00000  52.84884   49.94997   48.31149   47.61302   57.22762
#> 2005-Bush   52.84884   0.00000   60.84406   56.85948   57.41080   66.00000
#> 2009-Obama  49.94997  60.84406    0.00000   47.98958   54.91812   56.12486
#> 2013-Obama  48.31149  56.85948   47.98958    0.00000   53.73081   56.45352
#> 2017-Trump  47.61302  57.41080   54.91812   53.73081    0.00000   59.98333
#> 2021-Biden  57.22762  66.00000   56.12486   56.45352   59.98333    0.00000
as.list(tstat4)
#> $`2001-Bush`
#> 2021-Biden  2005-Bush 2009-Obama 2013-Obama 2017-Trump 
#>   57.22762   52.84884   49.94997   48.31149   47.61302 
#> 
#> $`2005-Bush`
#> 2021-Biden 2009-Obama 2017-Trump 2013-Obama  2001-Bush 
#>   66.00000   60.84406   57.41080   56.85948   52.84884 
#> 
#> $`2009-Obama`
#>  2005-Bush 2021-Biden 2017-Trump  2001-Bush 2013-Obama 
#>   60.84406   56.12486   54.91812   49.94997   47.98958 
#> 
#> $`2013-Obama`
#>  2005-Bush 2021-Biden 2017-Trump  2001-Bush 2009-Obama 
#>   56.85948   56.45352   53.73081   48.31149   47.98958 
#> 
#> $`2017-Trump`
#> 2021-Biden  2005-Bush 2009-Obama 2013-Obama  2001-Bush 
#>   59.98333   57.41080   54.91812   53.73081   47.61302 
#> 
#> $`2021-Biden`
#>  2005-Bush 2017-Trump  2001-Bush 2013-Obama 2009-Obama 
#>   66.00000   59.98333   57.22762   56.45352   56.12486 
#> 
as.dist(tstat4)
#>            2001-Bush 2005-Bush 2009-Obama 2013-Obama 2017-Trump
#> 2005-Bush   52.84884                                           
#> 2009-Obama  49.94997  60.84406                                 
#> 2013-Obama  48.31149  56.85948   47.98958                      
#> 2017-Trump  47.61302  57.41080   54.91812   53.73081           
#> 2021-Biden  57.22762  66.00000   56.12486   56.45352   59.98333

# distances for specific documents
textstat_dist(dfmat, dfmat["2017-Trump", ], margin = "documents")
#> textstat_dist object; method = "euclidean"
#>            2017-Trump
#> 2001-Bush        47.6
#> 2005-Bush        57.4
#> 2009-Obama       54.9
#> 2013-Obama       53.7
#> 2017-Trump          0
#> 2021-Biden       60.0
(tstat5 <- textstat_dist(dfmat, dfmat[c("2009-Obama" , "2013-Obama"), ], margin = "documents"))
#> textstat_dist object; method = "euclidean"
#>            2009-Obama 2013-Obama
#> 2001-Bush        49.9       48.3
#> 2005-Bush        60.8       56.9
#> 2009-Obama          0       48.0
#> 2013-Obama       48.0          0
#> 2017-Trump       54.9       53.7
#> 2021-Biden       56.1       56.5
as.matrix(tstat5)
#>            2009-Obama 2013-Obama
#> 2001-Bush    49.94997   48.31149
#> 2005-Bush    60.84406   56.85948
#> 2009-Obama    0.00000   47.98958
#> 2013-Obama   47.98958    0.00000
#> 2017-Trump   54.91812   53.73081
#> 2021-Biden   56.12486   56.45352
as.list(tstat5)
#> $`2009-Obama`
#>  2005-Bush 2021-Biden 2017-Trump  2001-Bush 2013-Obama 
#>   60.84406   56.12486   54.91812   49.94997   47.98958 
#> 
#> $`2013-Obama`
#>  2005-Bush 2021-Biden 2017-Trump  2001-Bush 2009-Obama 
#>   56.85948   56.45352   53.73081   48.31149   47.98958 
#> 

if (FALSE) {
# plot a dendrogram after converting the object into distances
plot(hclust(as.dist(tstat4)))
}