Summarize documents as syntactic and lexical feature counts

Count syntactic and lexical features of documents such as tokens, types, sentences, and character categories.

textstat_summary(x, ...)

Arguments

x

corpus to be summarized

...

additional arguments passed through to dfm()

Details

Count the total number of characters, tokens and sentences as well as special tokens such as numbers, punctuation marks, symbols, tags and emojis.

chars = number of characters; equal to nchar()

sents = number of sentences; equal ntoken(tokens(x), what = "sentence")

tokens = number of tokens; equal to ntoken()

types = number of unique tokens; equal to ntype()

puncts = number of punctuation marks (^\p{P}+$)

numbers = number of numeric tokens (^\p{Sc}{0,1}\p{N}+([.,]*\p{N})*\p{Sc}{0,1}$)

symbols = number of symbols (^\p{S}$)

tags = number of tags; sum of pattern_username and pattern_hashtag in quanteda::quanteda_options()

emojis = number of emojis (^\p{Emoji_Presentation}+$)

Examples

if (Sys.info()["sysname"] != "SunOS") {
library("quanteda")
corp <- data_corpus_inaugural[1:5]
textstat_summary(corp)
toks <- tokens(corp)
textstat_summary(toks)
dfmat <- dfm(toks)
textstat_summary(dfmat)
}
#>          document chars sents tokens types puncts numbers symbols urls tags
#> 1 1789-Washington    NA    NA   1537   603    107       0       0    0    0
#> 2 1793-Washington    NA    NA    147    95     12       0       0    0    0
#> 3      1797-Adams    NA    NA   2577   801    259       0       0    0    0
#> 4  1801-Jefferson    NA    NA   1923   687    197       0       0    0    0
#> 5  1805-Jefferson    NA    NA   2380   781    214       0       0    0    0
#>   emojis
#> 1      0
#> 2      0
#> 3      0
#> 4      0
#> 5      0